Move and rename ResourceLoader
