+++ platform-sdk/consensus-hashgraph-impl/src/test/java/org/hiero/consensus/hashgraph/impl/consensus/MaxRoundCreatedTest.java
@@ -10,11 +10,11 @@ import com.hedera.pbj.runtime.ParseException;
 import com.hedera.pbj.runtime.io.stream.ReadableStreamingData;
 import com.swirlds.config.api.Configuration;
 import com.swirlds.config.extensions.test.fixtures.TestConfigBuilder;
-import com.swirlds.platform.test.fixtures.resource.ResourceLoader;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import org.hiero.base.utility.test.fixtures.io.ResourceExtractor;
 import org.hiero.consensus.hashgraph.impl.test.fixtures.consensus.ConsensusOutput;
 import org.hiero.consensus.hashgraph.impl.test.fixtures.consensus.TestIntake;
 import org.hiero.consensus.io.IOIterator;
@@ -39,7 +39,7 @@ public class MaxRoundCreatedTest {
 
     @BeforeEach
     void setup() throws IOException {
-        final ResourceLoader<MaxRoundCreatedTest> loader = new ResourceLoader<>(MaxRoundCreatedTest.class);
+        final ResourceExtractor<MaxRoundCreatedTest> loader = new ResourceExtractor<>(MaxRoundCreatedTest.class);
         final Path tempDir = loader.loadDirectory(RESOURCE_DIR);
         Files.move(tempDir, testDataDirectory, REPLACE_EXISTING);
     }
